pyproject.toml
[tool.poetry]
name = "discordpy-music-bot"
version = "0.1.0"
description = ""
authors = ["Moha-AlHanbali <moha.alhanbali@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
python-dotenv = "^0.19.2"
youtube_dl = "^2021.12.17"
PyNaCl = "^1.4.0"
ffmpeg = "^1.4"
static-ffmpeg = "^1.0.9"

[tool.poetry.dev-dependencies]
black = "^21.12b0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
